rename 'utils' package to 'util'
