+++ src/main/java/com/karankumar/bookproject/backend/util/DateUtils.java
@@ -15,7 +15,7 @@
     If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.karankumar.bookproject.backend.utils;
+package com.karankumar.bookproject.backend.util;
 
 import com.helger.commons.annotation.VisibleForTesting;
