+++ src/main/java/com/karankumar/bookproject/ui/shelf/component/BookGridColumn.java
@@ -18,7 +18,7 @@
 package com.karankumar.bookproject.ui.shelf.component;
 
 import com.karankumar.bookproject.backend.entity.Book;
-import com.karankumar.bookproject.backend.utils.BookUtils;
+import com.karankumar.bookproject.backend.util.BookUtils;
 import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.data.renderer.LocalDateRenderer;
 import com.vaadin.flow.function.ValueProvider;
