+++ src/test/java/com/karankumar/bookproject/backend/service/ReadingGoalServiceTest.java
@@ -24,7 +24,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import static com.karankumar.bookproject.utils.ReadingGoalTestUtils.resetGoalService;
+import static com.karankumar.bookproject.util.ReadingGoalTestUtils.resetGoalService;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assumptions.assumeThat;
