+++ src/test/java/com/karankumar/bookproject/backend/statistics/util/StatisticTestUtils.java
@@ -15,7 +15,7 @@
     If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.karankumar.bookproject.backend.statistics.utils;
+package com.karankumar.bookproject.backend.statistics.util;
 
 import com.karankumar.bookproject.backend.entity.Author;
 import com.karankumar.bookproject.backend.entity.Book;
