+++ src/test/java/com/karankumar/bookproject/backend/util/CustomShelfUtilsTest.java
@@ -1,4 +1,4 @@
-package com.karankumar.bookproject.backend.utils;
+package com.karankumar.bookproject.backend.util;
 
 import com.karankumar.bookproject.annotations.IntegrationTest;
 import com.karankumar.bookproject.backend.entity.Author;
